pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai-compliance"
version = "1.0.0"
description = "Multi-agent AI pipeline for automated FinCEN SAR, CTR, and OFAC rejection report generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.80.0",
    "crewai[openai]>=0.80.0",
    "pydantic>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "streamlit>=1.33.0",
    "openai>=1.30.0",
    "pypdf>=4.0.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "weaviate-client>=3.24.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "streamlit_app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests"]
